Don't override a package if all versions are deprecated
Consider

  active-repositories:
    , repo-a
    , repo-b:override

and a package mypkg of which all versions in repo-b are deprecated (i.e. none is
preferred).

Before this patch, the solver would ignore all versions of mypkg in repo-a. After
this patch, it considers versions of mypkg in both repo-a and repo-b.

See https://github.com/haskell/cabal/issues/8502 for motivation